+++ src/entity/artwork/artwork.args.ts
@@ -1,5 +1,5 @@
 import { ArtworkImageType, ListType } from '../../types/enums.js';
-import { ArgsType, Field, Float, ID, InputType, Int } from 'type-graphql';
+import { ArgsType, Field, ID, InputType, Int } from 'type-graphql';
 import { DefaultOrderArgs, PaginatedFilterArgs } from '../base/base.args.js';
 import { examples } from '../../modules/engine/rest/example.consts.js';
 import { ObjParamsType } from '../../modules/rest/decorators/ObjParamsType.js';
@@ -81,7 +81,7 @@ export class ArtworkFilterArgs {
 	@ObjField(() => [ArtworkImageType], { nullable: true, description: 'filter by Artwork Image Types', example: ArtworkImageType.front })
 	types?: Array<ArtworkImageType>;
 
-	@Field(() => Float, { nullable: true })
+	@Field(() => Int, { nullable: true })
 	@ObjField({ nullable: true, description: 'filter by Creation timestamp', min: 0, example: examples.timestamp })
 	since?: number;
